Offer download of lodgeit.py with link to plain file on cgit
The current link is to the 'tree' version.  If the user right-clicks
and selects "save as", the file will be downloaded as an HTML file.

If the user clicks through to the tree page, another click is needed
on the "plain" link to either download the file with "save link as"
or to navigate to the plain link and then download.

Link directly to the "plain" page instead. This allows users to
directly download the file with "save link as".

Change-Id: I5d6a2f299ee264e5f06f68f72933a575b9042945